[refactor] PlaylistService DTO 조립/커서 로직 분리
